LayoutTests/storage/indexeddb/modern/deletedatabase-request-expected.txt

This test calls deleteDatabase on window.indexedDB and verifies that an IDBOpenDBRequest object is returned.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: [object IDBOpenDBRequest]
PASS successfullyParsed is true

TEST COMPLETE
